Like,Cart jwt add

+++ controller/LikeController.js
@@ -1,36 +1,76 @@
+const jwt = require("jsonwebtoken");
 const conn = require("../mariadb"); //db모듈
 const { StatusCodes } = require("http-status-codes");
-
+const dotenv = require("dotenv");
+dotenv.config();
 //좋아요 추가
 const addLike = (req, res) => {
   const { id } = req.params;
   likes_book_id = id;
-  const { user_id } = req.body;
-  let sql = "INSERT INTO likes (user_id, liked_book_id) Values(?,?)";
-  let values = [user_id, likes_book_id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
+  // const { user_id } = req.body; => 수정 jwt이용해서 가져온다.
+
+  let authorization = ensureAuthorization(req, res);
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다 다시 로그인 하세요",
+    });
+  } else {
+    let sql = "INSERT INTO likes (user_id, liked_book_id) Values(?,?)";
+    let values = [authorization.id, likes_book_id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const removeLike = (req, res) => {
   const { id } = req.params;
   likes_book_id = id;
-  const { user_id } = req.body;
-  let values = [user_id, likes_book_id];
-  let sql = "DELETE FROM likes WHERE user_id =? AND liked_book_id =?";
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
+
+  let authorization = ensureAuthorization(req, res);
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다 다시 로그인 하세요",
+    });
+  } else {
+    let values = [authorization.id, likes_book_id];
+    let sql = "DELETE FROM likes WHERE user_id =? AND liked_book_id =?";
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
+
+function ensureAuthorization(req, res) {
+  try {
+    let receivedJwt = req.headers["authorization"];
+    console.log("received jswt: ", receivedJwt);
+
+    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
+    console.log(decodedJwt);
+    return decodedJwt;
+  } catch (err) {
+    console.log(err.name);
+    console.log(err.message);
+    return err;
+  }
+}
 
 module.exports = {
   addLike,
